+++ CHANGELOG.md
@@ -86,6 +86,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Moved listeners/get to analytics/listeners [Issue 71](https://github.com/Lovinity/wwsu/issues/71)
  - Renamed songs/queue-add to songs/queue-top-add [Issue 67](https://github.com/Lovinity/wwsu/issues/67)
  - Moved sails.models.attendance.createRecord to sails.helpers.attendance.createRecord. [Issue 12](https://github.com/Lovinity/wwsu/issues/12)
+ - Moved all functions in the calendar model for syncing events to sails.helpers.calendar.sync(ignoreChangingState = false). [Issue 12](https://github.com/Lovinity/wwsu/issues/12)
 
 ## [5.1.0] - 2019-03-15
 ### Deprecated